Revert some PROGMEM changes, partially functional
Some of the PROGMEM movements caused access violations.  Undo them and build
again.  Generates a recognizable speaking, but still slower than realtime and
with hiccups.

+++ libhelix-aac/dequant.c
@@ -295,7 +295,7 @@ int Dequantize(AACDecInfo *aacDecInfo, int ch)
 {
 	int gp, cb, sfb, win, width, nSamps, gbMask;
 	int *coef;
-	const /*short*/ int *sfbTab;
+	const short *sfbTab;
 	unsigned char *sfbCodeBook;
 	short *scaleFactors;
 	PSInfoBase *psi;
